ref: Reorganize Day 4 files
