+++ lib/puzzles/2023/day04/day04.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require_relative "part1"
+require_relative "part2"
+
+module AdventOfCode
+  module Puzzles2023
+    ##
+    # Advent of Code 2023 - Day 4
+    # https://adventofcode.com/2023/day/4
+    # {include:file:lib/puzzles/2023/day04/README.md}
+    module Day04
+    end
+  end
+end
