+++ lib/puzzles/2023/day04/input.txt
